RBAC in spring security

+++ src/main/java/com/example/journalApplication/service/JournalEntryService.java
@@ -49,8 +49,10 @@ public class JournalEntryService {
 
     public  void deleteById(ObjectId id , String userName){
         User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+        boolean removed=user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if(removed){
+            userService.saveEntry(user);
+            journalEntryRepository.deleteById(id);
+        }
     }
 }
